feat: image gpt - remove separate codebook

+++ src/transformer_vq_vae/train_image_gpt.py
@@ -396,7 +396,7 @@ def sample_images(
     image_gpt_copy = copy.deepcopy(image_gpt)
     image_gpt_copy = torch.compile(image_gpt_copy)
     image_gpt_copy.eval()
-    # image_gpt_copy.half()
+    image_gpt_copy.half()
 
     vq_vae_model.eval()
